Refactored the main clicker area so now cards are easily exchanged in and out for scaleability

+++ src/com/tbonegames/cards/Cards.java
@@ -11,7 +11,7 @@ public class Cards {
 	//this is used so there won't be duplicates when drafting in the future. ID 1, 2, 3, etc. Also, this is useful for counting the amount of items in stuff in game.
 	public int idNumber;
 	
-	public int clickerIncrease;
+	public double clickerIncrease;
 	
 	public int defenseBonus;
 	
@@ -35,6 +35,5 @@ public class Cards {
 	
 	public int cardPrice;
 	public int cardPriceIncrease;
-	
 	
 }
